2.0.20180516.1

+++ library/src/main/java/name/zeno/android/behavior/VerticalScrollingBehavior.kt
@@ -37,7 +37,7 @@ abstract class VerticalScrollingBehavior<V : View> : CoordinatorLayout.Behavior<
 
   constructor(context: Context, attrs: AttributeSet) : super(context, attrs) {}
 
-  @IntDef(SCROLL_DIRECTION_UP.toLong(), SCROLL_DIRECTION_DOWN.toLong())
+  @IntDef(SCROLL_DIRECTION_UP, SCROLL_DIRECTION_DOWN)
   @Retention(AnnotationRetention.SOURCE)
   annotation class ScrollDirection()
